rename weback.config
